+++ hive/example/lib/hive_registrar.g.dart
@@ -1,0 +1,8 @@
+import 'package:hive_ce/hive.dart';
+import 'package:example/main.dart';
+
+extension HiveRegistrar on HiveInterface {
+  void registerAdapters() {
+    registerAdapter(PersonAdapter());
+  }
+}