+++ hive_generator/example/lib/hive_registrar.g.dart
@@ -1,0 +1,12 @@
+import 'package:hive_ce/hive.dart';
+import 'package:example/types.dart';
+
+extension HiveRegistrar on HiveInterface {
+  void registerAdapters() {
+    registerAdapter(Class1Adapter());
+    registerAdapter(Class2Adapter());
+    registerAdapter(EmptyClassAdapter());
+    registerAdapter(IterableClassAdapter());
+    registerAdapter(Enum1Adapter());
+  }
+}