+++ hive_generator/lib/src/helper.dart
@@ -36,3 +36,36 @@ void check(bool condition, Object error) {
     throw error;
   }
 }
+
+/// Returns [element] as [InterfaceElement] if it is a class or enum
+InterfaceElement getClass(Element element) {
+  check(
+    element.kind == ElementKind.CLASS || element.kind == ElementKind.ENUM,
+    'Only classes or enums are allowed to be annotated with @HiveType.',
+  );
+
+  return element as InterfaceElement;
+}
+
+/// Generate a default adapter name from the type name
+String generateAdapterName(String typeName) {
+  var adapterName =
+      '${typeName}Adapter'.replaceAll(RegExp(r'[^A-Za-z0-9]+'), '');
+  if (adapterName.startsWith('_')) {
+    adapterName = adapterName.substring(1);
+  }
+  if (adapterName.startsWith(r'$')) {
+    adapterName = adapterName.substring(1);
+  }
+  return adapterName;
+}
+
+/// Get the adapter name from the annotation or generate a default name
+String getAdapterName(String typeName, ConstantReader annotation) {
+  final annAdapterName = annotation.read('adapterName');
+  if (annAdapterName.isNull) {
+    return generateAdapterName(typeName);
+  } else {
+    return annAdapterName.stringValue;
+  }
+}